refactor: 提取图片裁剪逻辑为通用组件
